rename presentation.md --> introduction.md
In English, "presentation" usually refers to a particular object or event related to presenting an idea to an audience (say, a slideshow presentation, or a public session where an announcement is made) rather than the actual content of such an action. For websites, the term most likely to convey the intended meaning here is "about", but that's already taken by another page (which should IMO be called "credits", but that's already taken by *yet another* page!), so as a second-best option, I'm proposing "introduction" instead.